+++ CHANGELOG.md
@@ -1,6 +1,15 @@
+# v0.3.0
+
+- added `unsaved-check` to `run` command
+- upgraded `render_constants`
+- added `check-unsaved` hook
+- fixed `run outjob` return code
+
 # v0.2.2
+
 - fixed encoding issues
 - fixed project parameters parsing
+
 # v0.2.1
 
 - updated pre-commit config generation
@@ -8,7 +17,7 @@
 # v0.2.0
 
 - fixed bugs with execution and improved error handling
-- upgraded run command to allow for output job files execution 
+- upgraded run command to allow for output job files execution
 
 # v0.1.2